% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellam.R
\name{dskellam}
\alias{dskellam}
\title{PMF of Skellam distribution i.e. difference of two independent Poisson RVs}
\usage{
dskellam(k, lambda, mu)
}
\arguments{
\item{k}{argument of PMF}

\item{lambda}{mean number of points of favorite team}

\item{mu}{mean number of points of underdog team}
}
\value{
numeric or vector
}
\description{
{The Skellam distribution for the difference of two independent
Poisson RVs is used to model point-spread in games with goals worth unit-points.}
}
