% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryKelly.R
\name{entropy_binary}
\alias{entropy_binary}
\title{Growth rate for Kelly fraction for arbitrary binary (non-random) payoffs}
\usage{
entropy_binary(a, b, p)
}
\arguments{
\item{a}{the amount won on top of the wager}

\item{b}{the wager lost}

\item{p}{the probability of winning}
}
\value{
numeric
}
\description{
{The growth rate or entropy of the Kelly strategy}

{The average log wealth at the optimal strategy.}
}
