% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{entropy_moneyline}
\alias{entropy_moneyline}
\title{Entropy-growth for log-optimal moneyline bets}
\usage{
entropy_moneyline(p, a, b, u, v)
}
\arguments{
\item{p}{true probability of favorite winning}

\item{a}{winnings if favorite wins}

\item{b}{wager to throw down for favorite \eqn{-b}}

\item{u}{winnings if underdog wins \eqn{+u}}

\item{v}{wager to throw down for underdog}
}
\value{
numeric
}
\description{
{The entropy-growth for the log-optimal bet on two wagers for favorite
vs underdog team to win.}
}
