% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_check.R
\name{entropyBinary}
\alias{entropyBinary}
\title{Growth rate of the Kelly-criterion for binary IID bets}
\usage{
entropyBinary(p, a, b)
}
\arguments{
\item{p}{probability of success}

\item{a}{winnings on top of wager}

\item{b}{wager}
}
\value{
numeric
}
\description{
Growth rate of the Kelly-criterion for binary IID bets
}
