% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liveLinesScraper.R
\name{espn_nfl_line}
\alias{espn_nfl_line}
\title{Scrape NFL line data from ESPN}
\usage{
espn_nfl_line()
}
\value{
data.frame
}
\description{
{Pulls and formats data from ESPN for current schedules line.}
}
