% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperNFL.R
\name{espn_nfl_scrape}
\alias{espn_nfl_scrape}
\title{ESPN stats scraper}
\usage{
espn_nfl_scrape(team)
}
\arguments{
\item{team}{team name, usually first three letters of city or team-name.}
}
\value{
list
}
\description{
{Get mean number of touchdowns and (good) field goals per game for the
current season}
}
