% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalNFL.R
\name{growth_rate}
\alias{growth_rate}
\title{Optimal allocations among many independent bets}
\usage{
growth_rate(x, ps, a, b)
}
\arguments{
\item{x}{the allocation for the bets}

\item{ps}{vector of chances}

\item{a}{vector of win payouts}

\item{b}{vector of risk payins}
}
\value{
vector
}
\description{
{Compute the growth-rate for a given allocation under the
independent simultaneous binary wager model.}
}
