% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryKelly.R
\name{kelly_binary}
\alias{kelly_binary}
\title{Kelly fraction for arbitrary binary (non-random) payoffs}
\usage{
kelly_binary(p, a, b)
}
\arguments{
\item{p}{the probability of winning}

\item{a}{the amount won on top of the wager}

\item{b}{the wager lost}
}
\value{
numeric
}
\description{
{The classical Kelly criterion formula for Bernoulli bets of arbitrary odds.}

{The formula is well known, i.e. on wikipedia.}
}
