% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_check.R
\name{kellyBinary}
\alias{kellyBinary}
\title{Kelly-criterion for binary IID bets}
\usage{
kellyBinary(p, a, b)
}
\arguments{
\item{p}{probability of success}

\item{a}{winnings on top of wager}

\item{b}{wager}
}
\value{
numeric
}
\description{
Kelly-criterion for binary IID bets
}
