% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalNFL.R
\name{kellyMultipleBets}
\alias{kellyMultipleBets}
\title{Optimal allocations among many independent bets}
\usage{
kellyMultipleBets(ps, a, b, restraint = 1)
}
\arguments{
\item{ps}{vector of chances}

\item{a}{vector of win payouts}

\item{b}{vector of risk payins}

\item{restraint}{percentage of wealth to restrict to}
}
\value{
vector
}
\description{
{Given independent alternatives to wager on on the same trial, going up to
a percentage of one's bankroll, optimal allocations are computed via
log-maximization of terminal wealth.}
}
