% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{log_optimal_moneyline_bet}
\alias{log_optimal_moneyline_bet}
\title{Kelly-criterion for moneyline bets}
\usage{
log_optimal_moneyline_bet(teams, fav_minus, underdog_plus, n = 10^5)
}
\arguments{
\item{teams}{vector of team names}

\item{fav_minus}{risk for betting on favorite (-b)}

\item{underdog_plus}{profit for betting on underdog (+a)}

\item{n}{number of variates to use in Monte-Carlo estimations}
}
\value{
list
}
\description{
{Log optimal wager strategy for moneyline betting. Based on
linear combination of Poisson RVs model for points.}
}
