% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{log_optimal_spread_bet}
\alias{log_optimal_spread_bet}
\title{Kelly-criterion for moneyline bets}
\usage{
log_optimal_spread_bet(teams, spread, n = 10^5)
}
\arguments{
\item{teams}{vector of team names}

\item{spread}{the point-spread}

\item{n}{number of variates to use in Monte-Carlo estimations}
}
\value{
list
}
\description{
{Log optimal wager strategy for moneyline betting. Based on
linear combination of Poisson RVs model for points.}
}
