% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{log_optimal_total_bet}
\alias{log_optimal_total_bet}
\title{Kelly-criterion for Total points bets}
\usage{
log_optimal_total_bet(teams, line, total_wager = 110, n = 10^5)
}
\arguments{
\item{teams}{vector of team names}

\item{line}{the line for betting over/under}

\item{total_wager}{wager risked (-b)}

\item{n}{number of variates to use in Monte-Carlo estimations}
}
\value{
list
}
\description{
{Log optimal wager strategy for total points betting. Based on
linear combination of Poisson RVs model for points.}
}
