% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{logOptimalNFL}
\alias{logOptimalNFL}
\title{Log optimal allocations among independent bets for NFL}
\usage{
logOptimalNFL(
  bankroll = 444,
  estimates = NULL,
  restraint = 1,
  wager = 110,
  top = 2
)
}
\arguments{
\item{bankroll}{wealth to bet with}

\item{estimates}{object returned from \code{nfl_model_chances}}

\item{restraint}{percentage of wealth to use}

\item{wager}{wager to risk for total over/under (110 default)}

\item{top}{how many games to bet on}
}
\value{
list of four data.frames \code{over} and \code{under} and \code{fave} and \code{underdog}
both containing the same column names except the last one.
Contained is the model-estimate for the outcome together with the
optimal fraction of wealth to bet.
}
\description{
{Kelly-criterion applied to multiple independent bets per
trial.}
}
