% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{logOptimalTotals}
\alias{logOptimalTotals}
\title{Log optimal allocations among independent bets}
\usage{
logOptimalTotals(
  bankroll = 444,
  tdat = NULL,
  n = 5 * 10^4,
  restraint = 1,
  wager = 110,
  top = 2
)
}
\arguments{
\item{bankroll}{wealth to bet with}

\item{tdat}{scraped data from \code{espn_live_line}}

\item{n}{number of trials in MC estimations}

\item{restraint}{percentage of wealth to use}

\item{wager}{wager to risk}

\item{top}{how many games to bet on}
}
\value{
list of two data.frames \code{over} and \code{under}
both containing the same column names except the last one.
Contained is the model-estimate for the outcome together with the
optimal fraction of wealth to bet.
}
\description{
{Kelly-criterion applied to multiple independent bets per
trial.}
}
