% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{logOptimalWeekBet}
\alias{logOptimalWeekBet}
\title{Log optimal bet for an independent but not identically distributed sequence}
\usage{
logOptimalWeekBet(chances, a = 100/110, b = 1)
}
\arguments{
\item{chances}{sequence of chances of winning the bet}

\item{a}{payout, default 100}

\item{b}{risk, default 110}
}
\value{
numeric
}
\description{
{Log optimal growth for a finite sequence of bets with different
probabilities.}
}
