% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{logOptWeekTotal}
\alias{logOptWeekTotal}
\title{Maximize log-growth of a sequence of bets with different chances}
\usage{
logOptWeekTotal(tdat = NULL, n = 5 * 10^4)
}
\arguments{
\item{tdat}{data returned from \code{espn_nfl_line}}

\item{n}{number of variates}
}
\value{
list
}
\description{
{Work in progress model for non-identical chance bets.}
}
