% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagers.R
\name{moneyline}
\alias{moneyline}
\title{The chance of the favorite winning}
\usage{
moneyline(lambda, mu)
}
\arguments{
\item{lambda}{mean number of points for the favorite \eqn{X}}

\item{mu}{mean number of points for the underdog \eqn{Y}}
}
\value{
numeric or vector
}
\description{
{The probability of the event \eqn{\{X-Y>0\}}}
}
