% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{moneyline_analysis}
\alias{moneyline_analysis}
\title{Model analysis for a moneyline bet}
\usage{
moneyline_analysis(bankroll, favorite, underdog, lambda, mu, trials = 30)
}
\arguments{
\item{bankroll}{the amount to bet with}

\item{favorite}{the \eqn{-b} in \eqn{a} to \eqn{b} odds for the favorite.}

\item{underdog}{the \eqn{+u} in \eqn{u} to \eqn{v} odds for the underdog.}

\item{lambda}{mean number of points per game by favorite}

\item{mu}{mean number of points by per game by underdog}

\item{trials}{number of trials to simulate}
}
\value{
list of odds and decision rule
}
\description{
{Log optimal strategy for gambling on moneyline event
under a Skellam distribution for the point spread.}
}
