% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_endpoints.R
\name{nba_endpoints}
\alias{nba_endpoints}
\title{Convert team name string to team endpoint for ESPN url}
\usage{
nba_endpoints(teamName)
}
\arguments{
\item{teamName}{the full team name e.g. "Boston Bruins" etc.}
}
\value{
string
}
\description{
{Backend function for converting team name from string to url endpoint.}
}
\details{
{The statistics page url for an individual team on ESPN ends with a three letter string
resembling the city or team name of a given team. For example "Arizona Cardinals" has the endpoint
"ari". Exceptions include
\itemize{
\item Green Bay Packers is "gb"
\item JacksonVille Jaguars is "jax"
\item Kansas City Chiefs is "kc", etc.}}
}
