% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_check.R
\name{nba_over}
\alias{nba_over}
\title{Compute optimal over allocations for NBA games}
\usage{
nba_over(
  bankroll,
  underdogs,
  faves,
  lines,
  b,
  n = 1000,
  alpha = 0.15,
  ci = TRUE,
  method = "cf"
)
}
\arguments{
\item{bankroll}{the account value to bet with}

\item{underdogs}{list of underdog team names}

\item{faves}{list of favorite team names}

\item{lines}{vector of lines}

\item{b}{vector of wagers (typically 110)}

\item{n}{number of steps to use in CF integration or MC trials}

\item{alpha}{significance level for LBs and UBs for estimates}

\item{ci}{boolean for using CI}

\item{method}{string "cf" or "mc" for characteristic function method or Monte-Carlo}
}
\value{
prints out the bets.
}
\description{
{Wrapper to computing over chances and allocaitons given series of games.}
}
