% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_scrapers.R
\name{nba_stats}
\alias{nba_stats}
\title{ESPN stats scraper for NBA data}
\usage{
nba_stats(team)
}
\arguments{
\item{team}{team name, the full name}
}
\value{
Returns a vector of mean 2pters, 3pters, and free-throws as well as number of games played
}
\description{
{Get mean number of goals per game for a team.}
}
