% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_kelly.R
\name{nfl_kelly}
\alias{nfl_kelly}
\title{Compute log-optimal bets for NFL wagers}
\usage{
nfl_kelly(tdat = NULL, n = 5 * 10^4)
}
\arguments{
\item{tdat}{data returned from \code{espn_nfl_line()}}

\item{n}{number of simulations}
}
\value{
list of the ESPN data data.frame, optimal allocations under Kelly-Criterion, and growth-rates
}
\description{
{Compute bets that maximize log wealth for NFL wagers from ESPN
schedule.}
}
