% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNFL.R
\name{nfl_model_chances}
\alias{nfl_model_chances}
\title{Estimate model-chances of moneyline, point-spread, and o/u events}
\usage{
nfl_model_chances(tdat = NULL, n = 5 * 10^4)
}
\arguments{
\item{tdat}{data returned from \code{espn_nfl_line()}}

\item{n}{number of simulations}
}
\value{
list of model-estimated chances for outcomes in moneyline, point-spread, and o/u
}
\description{
{Using a linear combination of Poisson RVs to model the points of a given
team, independently, this function computes estimates of chances of moneyline, point-spread,
and o/u events via Monte-Carlo averaging.}
}
