% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_models.R
\name{nfl_moneyline_cdf}
\alias{nfl_moneyline_cdf}
\title{Monte-Carlo estimates of moneyline distribution}
\usage{
nfl_moneyline_cdf(fav_means, underdog_means, n = 10^6)
}
\arguments{
\item{fav_means}{vector of mean number of touchdowns and (good) field goals of favorite team}

\item{underdog_means}{vector of mean number of touchdowns and (good) field goals of underdog team}

\item{n}{number of variates to use in Monte-Carlo estimate}
}
\value{
data.frame of favorite win chance and underdog win(+tie) chance.
}
\description{
{Compute via Monte-Carlo simulations the probability of money-line events
for a given favorite and underdog team. See the details documentation for the total cdf for
more details on the model.}
}
