% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_scrapers.R
\name{nfl_stats}
\alias{nfl_stats}
\title{ESPN stats scraper}
\usage{
nfl_stats(team)
}
\arguments{
\item{team}{team name, the full name}
}
\value{
a named list containing
\itemize{
\item \code{scraped} a list containing the scraped data of
\code{record}, \code{total_tds}, and \code{fg_stats}
\item \code{means} }
}
\description{
{Get mean number of touchdowns and (good) field goals per game for the
current season}
}
