% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfsNFL.R
\name{nfl_total_cdf}
\alias{nfl_total_cdf}
\title{Monte-Carlo estimates of Total points distribution}
\usage{
nfl_total_cdf(line, means, engine = "cf", n = 10^6)
}
\arguments{
\item{line}{the line}

\item{means}{the vector of mean number of touchdowns and (good) field goals
of both teams cumulatively}

\item{engine}{"cf" or "mc"}

\item{n}{number of variates to use in Monte-Carlo estimation.}
}
\value{
data.frame of over, under, and push chances.
}
\description{
{Compute via Monte-Carlo probabilities of being over or under a given line.}
}
\details{
{We model the point process of a football team by a linear combination of Poisson RVs,
where the coefficients are the points earned for each type of goal. Since the data we use for the
mean number of touchdowns includes all touchdowns not just ones that have no conversion, the coefficients for the
the conversion points will be 1 and 2 rather than 7 and 8. I am not a football expert, so contact me if you have better ideas.}
}
