% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_scrapers.R
\name{nhl_stats}
\alias{nhl_stats}
\title{ESPN stats scraper for NHL data}
\usage{
nhl_stats(teamName)
}
\arguments{
\item{teamName}{team name, the full name}
}
\value{
a named list containing
\itemize{
\item \code{scraped} a list containing the scraped data of
\code{record}, \code{total_tds}, and \code{fg_stats}
\item \code{means} }
}
\description{
{Get mean number of goals per game for a team.}
}
