% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_nhl.R
\name{nhl_stats_endpoint}
\alias{nhl_stats_endpoint}
\title{Endpoints for NHL teams on ESPN.com}
\usage{
nhl_stats_endpoint(teamName)
}
\arguments{
\item{teamName}{the full team name, e.g. "Boston Bruins"}
}
\value{
string, the url endpoint for the team's statistics.
}
\description{
{Returns the appropriate endpoint for a NHL team. Internal use only.}
}
