% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagers.R
\name{overunder}
\alias{overunder}
\title{The chance of a given point spread under the Skellam model}
\usage{
overunder(k, lambda, mu, over = TRUE)
}
\arguments{
\item{k}{the size of the point spread (in the direction of the favorite)}

\item{lambda}{mean number of points for the favorite}

\item{mu}{mean number of points for the underdog}

\item{over}{boolean, true for over and false for under}
}
\value{
numeric or vector
}
\description{
{The probability of the event \eqn{\{X+Y>k\}} or its complement.}
}
