% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearCombPoints.R
\name{phi_X}
\alias{phi_X}
\title{CF of a linear combination of Poisson RVs}
\usage{
phi_X(u, lambdas, points)
}
\arguments{
\item{u}{real number argument}

\item{lambdas}{mean rates}

\item{points}{coefficients of points per goal}
}
\value{
numeric
}
\description{
{CF of a linear combination of Poisson variables representing
total points in a game.}
}
