% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearCombPoints.R
\name{plcpois}
\alias{plcpois}
\title{The CDF of a linear combination of Poisson RVs}
\usage{
plcpois(l, lambdas, points, grid = 300, engine = "cf")
}
\arguments{
\item{l}{integer argument, must be positive}

\item{lambdas}{mean rates}

\item{points}{coefficients representing points per goal}

\item{grid}{grid size for numeric integration}

\item{engine}{either "cf" or "mc"}
}
\value{
numeric
}
\description{
{Numeric integration of CDF of a linear combination of Poisson
RVs via inverting the CF using trapezoid rule.}
}
