% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagers.R
\name{point_spread}
\alias{point_spread}
\title{The chance of a given point spread under the Skellam model}
\usage{
point_spread(k, lambda, mu)
}
\arguments{
\item{k}{the size of the point spread (in the direction of the favorite)}

\item{lambda}{mean number of points for the favorite}

\item{mu}{mean number of points for the underdog}
}
\value{
numeric or vector
}
\description{
{The probability of the event \eqn{\{X-Y>k\}}}
}
