% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellam.R
\name{rskellam}
\alias{rskellam}
\title{Simulate variates of Skellam distribution i.e. difference of two independent Poisson RVs}
\usage{
rskellam(n, lambda, mu)
}
\arguments{
\item{n}{number of variates to sample from the distribution}

\item{lambda}{mean number of points of favorite team}

\item{mu}{mean number of points of underdog team}
}
\value{
numeric or vector
}
\description{
{The Skellam distribution for the difference of two independent
Poisson RVs is used to model point-spread in games with goals worth unit-points.}
}
