% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrapeStatsNFL}
\alias{scrapeStatsNFL}
\title{ESPN stats scraper}
\usage{
scrapeStatsNFL(team)
}
\arguments{
\item{team}{team name, usually first three letters of city or team-name.}
}
\value{
a named list containing
\itemize{
\item \code{scraped} a list containing the scraped data of
\code{record}, \code{total_tds}, and \code{fg_stats}
\item \code{means} }
}
\description{
{Get mean number of touchdowns and (good) field goals per game for the
current season}
}
