% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sim_moneyline}
\alias{sim_moneyline}
\title{Simulate a series of IID trials of moneyline under log-optimal growth}
\usage{
sim_moneyline(bankroll, p, wagers, trials = 100)
}
\arguments{
\item{bankroll}{initial bankroll to bet}

\item{p}{true chance of outcome}

\item{wagers}{the vector of wagers, see details}

\item{trials}{the number of trials to simulate}
}
\value{
vector
}
\description{
{Simulate a series of moneyline trials under log-optimal growth.}
}
\details{
{The argument \code{wagers} must contain four wagers, the first two the odds
for the favorite, the latter two the odds for the underdog.}
}
