% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{team_endpoint}
\alias{team_endpoint}
\title{Convert team name to team endpoint for ESPN url}
\usage{
team_endpoint(team_name)
}
\arguments{
\item{team_name}{team name e.g. "Arizona Cardinals" etc.}
}
\value{
string
}
\description{
{Backend function for converting team name from string to url endpoint.}
}
