% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_points.R
\name{total_points_chances}
\alias{total_points_chances}
\title{Model estimates of Total points distribution for two-teams}
\usage{
total_points_chances(line, means, points, method = "cf", n = 10^3)
}
\arguments{
\item{line}{the line}

\item{means}{vector of the team-wise sum of the mean number of goals of each type}

\item{points}{the coefficients of points earned per goal, i.e. in basketball (2,3,1) (see details for orders)}

\item{method}{"cf" or "mc"}

\item{n}{number of variates to use in Monte-Carlo estimation.}
}
\value{
data.frame of over, under, and push chances.
}
\description{
{Compute via Monte-Carlo probabilities of being over or under a given line.}
}
\details{
{We model the point process of a football team by a linear combination of Poisson RVs,
where the coefficients are the points earned for each type of goal. Since the data we use for the
mean number of touchdowns includes all touchdowns not just ones that have no conversion, the coefficients for the
the conversion points will be 1 and 2 rather than 7 and 8. I am not a football expert, so contact me if you have better ideas.}
}
