% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalNFL.R
\name{weekly_bets}
\alias{weekly_bets}
\title{Weekly bets for all categories of wagers}
\usage{
weekly_bets(bankroll, live_line = NULL, chances = NULL, wager = 110)
}
\arguments{
\item{bankroll}{initial wealth to bankroll}

\item{live_line}{the scraped line from ESPN}

\item{chances}{estimated model chances}

\item{wager}{the wager for over/under and spreads}
}
\value{
list
}
\description{
{Optimize allocations amongst weekly bets.
}
}
