% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_optimal.R
\name{weekTotalChances}
\alias{weekTotalChances}
\title{Compute chances of over/under for a given live line on NFL game week}
\usage{
weekTotalChances(tdat = NULL, n = 5 * 10^4)
}
\arguments{
\item{tdat}{data returned from \code{espn_nfl_line}}

\item{n}{number of variates}
}
\value{
list
}
\description{
{A linear combination of Poisson RVs is used to model the points of each team.}
}
