% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_nhl.R
\name{nhl_stats}
\alias{nhl_stats}
\title{Obtain the season stats for NHL teams by team name}
\usage{
nhl_stats(teamName)
}
\arguments{
\item{teamName}{string of team name, in full, e.g. "Boston Bruins"}
}
\value{
data.frame of the team's statistics
}
\description{
{Pull the ESPN data for NHL team's statistics.}
}
